refactor: refactored ticketType page to follow FillSubscription page
