+++ src/components/Payment/SelectingTicketType/ReserveOption.js
@@ -1,7 +1,7 @@
 import { Typography } from '@material-ui/core';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import Button from '../Form/Button';
+import Button from '../../Form/Button';
 
 export default function ReserveOption({ value, isRemote, includesHotel }) {
   const navigate = useNavigate();
