+++ src/components/Payment/SelectingTicketType/TicketMode.js
